first commit for 1.0
Collabtive 1.0
+ Improved Italian, Bulgarian, Hungarian and Brasilian
Portuguese translation
+ Added install-readme in Hungarian
+ Changed
name of the autoload function to cl_autoload and added it with
spl_autoload_register() to the autoload list, to not interfere with
other autoloaders
+ Implemented automatic update notification.
+
Completely re-written the PDF export. It now uses a custom subclass of
TCPDF to directly draw the tables for the reports from arrays instead of
using an HTML intermediate. This greatly speeds up PDF generation (esp
when many lines are exported) and also generates smaller PDFs. It should
also make customizing easier.
+ Fixed problem with missing "webcolor"
variable in TCPDF
+ Improved the Project log PDF Report by merging
fields, improving shading and adding automatic font stretching. This
makes for better readability.
+ Added new PDF report to the "my tasks"
block on the desktop. Now a user can export all his tasks to PDF.
+
Re-enabled permissions checking when deleting a timetracker entry. This
got disabled for testing purposes and was commited to trunk. Disabling
it posed a possible security risk.
+ Added permissions checking to
timetracker:add and timetracker:edit . Permissions set in the
useraccount where not properly enforced for the timetracker before.
+
Rewrote the settings functionality to work with a key/value table as the
datastore. This makes extensibility easier.