improve plugin system
to allow registering a custom template method for each template tag
registered
the first parameter for pluginManager::registerPlugin() now takes an
array of template tags.
This array contains another array for each entry, that contains a field
"tag" describing the tag to be registered and an optional field
"method".
The "method" field only needs to be implemented if the default template
method ::getTemplate is to be overriden.