Mitigate XSS attack on user profile page
See: https://devwerks.net/blog/16/how-not-to-use-html-purifier

I started by attempting a Smarty output filter that ran HTMLPurifier
over the entire rendered contents of page, which obviously did not
work well. The attack exploits the fact that there are two variables
escaped separately on input, but rendered on the same line on output.

This commit escapes the two user inputs together and assigns them to
a new template variable as a mitigation of this specific attack.